ar: initial alicecg2 implementation
Initial implementation of the ALICECG2 manifest format. The format
converts all images in a directory to another format and then archives
them. The converted images are stored on disk so that the transcode step
can be skipped when the converted image already exists and the source
image hasn't been modified.