ar extract: allow overriding table of contents
The --toc option can be used to provide a list of file names to use
instead of the archive's own table of contents. This is mainly useful
for AFAv3 archives where the table of contents is encrypted. It is
possible to extract the table of contents using Alicesoft's own
executables by compiling an ain file that runs the following code:

    int fp = TextFile.CreateWriter("out.txt");
    int c = AFAFactory.GetCountOfData(2); // '2' is CG archive type
    for (int i = 0; i < c; i++) {
        TextFile.WriteLine(fp, AFAFactory.GetTitleByIndex(2, i));
    }
    TextFile.Close(fp);