Use visitor pattern in .jaf compiler
Refactor every pass over the AST--except for the final compile pass--to
use the visitor pattern. This removes lots of duplicated AST navigation
code.

The type analysis pass on function calls is also significantly
simplified by the introduction of various "intermediate" types which are
only used in type analysis. For example, there is now a type AIN_SYSCALL
which is given to the callable "object" in code such as
"system.MsgBox(...)", and similar for HLL, method and builtin calls.

Also, v11+ builtin calls are now rewritten as HLL calls, so that the
regular HLL code paths can be reused.