Add `project build` command
Add a separate command for building .pje files, since this is no longer
an ain-specific operation.

Also, add the option to build archives as part of the .pje build
process, e.g.

    Archives = {
        "Rance10ModCG.manifest",
        "Rance10ModSound.manifest",
    }

(This isn't particularly well integrated, as it is necessary to
re-specify the output directory in each manifest file.)