Add `ex edit` command
This command appends data from a .txtex file to an existing .ex file.

When the -e option is specified, only the modified objects are written
to the output file (can be used to make mods which load a .ex file at
runtime).

When the -r option is specified, objects in the .ex file are replaced
rather than appended to.

These are very useful operations for modding as they allow a mod to
specify only the data to be added to the game, whereas previously a mod
would have to rebuild the entire .ex file.