pje: allow listing txtex files in source directory
txtex files can now be listed in Source/SystemSource within pje/inc
files.

If ExModName is given in the pje file, then these txtex files are
written to a separate ex file.

If ExInput and ExName are given (but not ExModName), the txtex files are
appended to the ExInput file's data.

If only ExName is given, then a completely new ex file is created from
the txtex files.

This will allow multiple mods to coexist within the same project
structure, each adding their own ex data.