pje: allow specifying archive dirs in .inc files
Directories containing files to be added to archives can now be
specified in .inc files. E.g.

    CG    = { "cg_src", "png", "cg_dst", "qnt" }
    Sound = { "sound_src", "ogg", "sound_src", "ogg" }
    Flat  = { "flat_src", "txtex", "flat_dst", "flat" }

The format is similar to the BATCHPACK manifest format. The .pje file
should specify the names of the archives to be written to the output
directory. E.g.

    CgName = "Rance10ModCG.afa"
    SoundName = "Rance10ModSound.afa"
    Flat = "Rance10ModFlat.afa"

This allows separate mods to add their own CGs, etc. while coexisting
within the same project structure.