pje: support directory wildcard
Strings with the format '*/file.ext' can now be used in
Source/SystemSource lists. This pattern will include all files named
'file.ext' in subdirectories of the current directory.

This is a bit of a hack to support a specific modding arrangement where
mods are installed into a project by copying a directory into the
SourceDir. The SourceDir would have a .inc file containing,

    Source { "*/source.inc" }

and then each mod directory would contain a file 'source.inc' which
defines the files to be included into the build.

This arrangement can be used to create a sort of modding framework that
would allow for games to be modded in a granular way.