asm: bytecode injection
Implement a system for injecting bytecode into a function at a known
offset.

This basically allows for creating mods that edit a function's bytecode
without distributing the entire function as part of the mod.

The only way to access this functionality is via the .pje build system.
In a Source or SystemSource list, a .jam file can be specified in the
following format,

    Source {
        "!inject!function_name!0x0FF5E7!code.jam",
    }

This will cause the bytecode in code.jam to be inserted at offset
0x0FF5E7 in the function `function_name`.