Refactor into 'core' and 'cli' directories
In order to support a GUI in the future, the source tree has been split
into a 'cli' directory, which contins CLI-specific code, and a 'core'
directory which contains code that could be used from either CLI or
GUI.

This mostly just moves the files around; the interfaces in 'core' are
still not really designed for GUI (they call `exit` on error, etc.).
Further work is needed here.