jaf: fix ref return types
The declarations pass is now split in two: a pass to register functions
and globals in the ain file, and a later pass to allocate local
variables for functions.

The latter pass needs to happen after type analysis, so that function
return types are known (because a local variable must be allocated for
calls that return ref types).