jaf: implement interface types and properties
Interface types are two-valued types consisting of a struct reference
and a vtable offset. For the most part they are treated the same as
scalar reference types. Calls to interface methods go through the
structs '<vtable>' member, which is always index 0 into the struct page.

For now it is not possible to define new interfaces or define structs
that implement interfaces. In order to use an existing interface, the
interface must be redefined so that the method signatures are known
(it's non-trivial to determine these automatically, because vtables are
initialized via bytecode in constructors).

Properties are just syntactic sugar for members with getter/setter
methods. For now only read/write to existing properties is supported;
there is no syntax for defining properties in a class.