jaf: add expect tests
These tests are generated by compiling a jaf file, and then dumping the
result to jam and json files.

The tests are run by recompiling the jaf file, and comparing the
resulting ain file to the ain file produced by combining the saved
jam/json files. If the ain files differ, the test fails.

This kind of test has 2 main benefits.
1. Changes to codegen are recorded in the git history.
2. No runtime is required to evaluate the test (meaning it can be run
   via CI for ain versions not supported by xsystem4).