Add `project init` command.
`alice project init <ain-file>` initializes a .pje project in the
current directory, targeting same VM version as the given .ain file.
This is mainly useful for setting up tests to see how a particular VM
behaves (i.e. for reverse engineering).

The `--mod` option can be given to initialize a mod project rather than
a blank project. This copies the .ain file, extracts the .ex file (if
present), and creates a .pje to build modified .ain/.ex files. Manifests
are also provided for packing archives, but the .pje entries to build
them are commented out initially.

This is essentially a translation of `extra/alice-mod-init.sh` into C.