jaf: implement option types
Option types are a two-valued type. The second value determines whether
an option is Some(value) or None.

Options can be unwrapped with `opt.Value(default)`. `IsSome()` and
`IsNone()` builtins are also implemented.

The `Value` builtin is based on observed bytecode patterns in Rance 10
and Dohna Dohna. `IsSome` and `IsNone` are not.

There is a difference in how options are represented in ain v12 and v14.
In v12, a value of -1 represents None (and the bytecode tests with
`== -1`), whereas in v14 a value of 1 represents None, and the bytecode
tests for `< 1` (Some) or `>= 1` (None).