jaf: improve enum support
When extending an enum, the default starting value is the highest value
in the existing enum + 1.

Enums can now be extended multiple times, and the declarations will be
combined before generating the enum's functions. Later declarations take
the earlier declaration's values into account when determining the
default starting value.

Together, this allows an enum to be extended multiple times and be
guaranteed to receive distinct values, without coordination.