ar: support conversion in ALICEPACK manifest
An optional dst format can now be specified for files in ALICEPACK
manifests. E.g.

    #ALICEPACK
    out.afa
    test.png,qnt
    test.x,ex

The src format is inferred from the filename, and the file extension of
the archived file will be changed to reflect the dst format. Unlike with
BATCKPACK, conversion happens in memory, so previously converted files
on disk cannot be used to speed up the process.

Only QNT and EX are supported as dst formats currently.