flat: add option to transcode images to png
Add `--png` option to `alice flat extract` and `--flat-png` option to
`alice ar extract`.

This not only transcodes the images, but puts the correct file name into
the generated .x file, so that the .flat will be rebuilt with .png
images instead of the original format. Rance 10 at least is able to use
such a flat file without issue, although it results in a ~4x increase in
the .afa file size to rebuild everything with pngs.