ar: add `--src-dir=` option for ALICEPACK manifest
Add an option to specify the source directory for files listed in an
ALICEPACK manifest (this option is specified on the first line of the
manifest, not the command line). This option is automatically included
in generated manifests when the `-o` option is given to `ar extract`.

This allows for the following pattern:

    alice ar extract -o flat --manifest=manifest.txt Rance10Flat.afa
    alice ar pack manifest.txt

Note that if the manifest is output to a directory other than the
current directory, this will fail because the `--src-dir` path of the
generated manifest is always relative to the current directory.