ar: implement caching for ALICEPACK manifests
When unpacking a CG archive with the `--manifest` option, a cache
directory is created and the raw files are extracted to this directory.
The `--cache-dir` manifest option is added to the generated manifest,
pointing at this directory.

When packing an archive where the `--cache-dir` manifest option was
given, converted files are drawn from the cache dir if the source file
timestamp is < the cached file timestamp.

This greatly improves the speed of repacking large CG archives.